+++ src/backend/typeDefs/mutations/relationalTypeDefs/pathwaysAndPerson/delete.js
@@ -16,6 +16,8 @@ const deletePathwaysAndPersonConnectionTypeDefs = gql`
     exclusionSettings: PathwaysAndPersonExclusionSettingsInput! # this type is on the upsert typeDefs
     startDate: String # must be string to preserve ISO short, then convert accounting for timezone in resolver
     endDate: String # must be string to preserve ISO short, then convert accounting for timezone in resolver
+    startQuarter: String # must be string to preserve ISO short, then convert accounting for timezone in resolver
+    endQuarter: String # must be string to preserve ISO short, then convert accounting for timezone in resolver
   }
 `