+++ src/frontend/components/BusinessObjectModal/PeopleModal/widgets/relational-widgets/OrganizationConnectionsWidget/alert-types.js
@@ -1,0 +1,3 @@
+export const INFLUENCER_ALERT_TYPE = 'Influencer'
+
+export const ALERT_TYPES = [INFLUENCER_ALERT_TYPE]